minor refactoring to clean up nrn_checkpoint.h
