+++ coreneuron/nrniv/nrn_checkpoint.h
@@ -32,9 +32,9 @@ THE POSSIBILITY OF SUCH DAMAGE.
  * functional protoype of writting checkpoint file in same format of input data
  * */
 
-#include "coreneuron/nrniv/nrn_filehandler.h"
-
 class NrnThread;
+class FileHandler;
+
 void write_checkpoint(NrnThread* nt,
                       int nb_threads,
                       const char* dir,
@@ -42,29 +42,6 @@ void write_checkpoint(NrnThread* nt,
 
 void checkpoint_restore_tqueue(NrnThread&, FileHandler&);
 
-template <typename T>
-T* chkpnt_soa2aos(T* data, int cnt, int sz, int layout, int* permute) {
-  // inverse of F -> data. Just a copy if layout=1. If SoA, original file order depends on
-  // padding and permutation.
-  // Good for a, b, area, v, diam, Memb_list.data, or anywhere values do not change.
-  T* d = new T[cnt * sz];
-  if (layout == 1) { /* AoS */
-    for (int i=0; i < cnt*sz; ++i) {
-      d[i] = data[i];
-    }
-  }else if (layout == 0) { /* SoA */
-    int align_cnt = nrn_soa_padded_size(cnt, layout);
-    for (int i=0; i < cnt; ++i) {
-      int ip = i;
-      if (permute) { ip = permute[i]; }
-      for (int j = 0; j < sz; ++j) {
-        d[i*sz + j] = data[ip + j*align_cnt];
-      }
-    }
-  }
-  return d;
-}
-     
 int* inverse_permute(int* p, int n);
 void nrn_inverse_i_layout(int i, int& icnt, int cnt, int& isz, int sz, int layout);
